\name{compound.Cox-package}
\alias{compound.Cox-package}
\alias{compound.Cox}
\docType{package}
\title{
Regression estimation based on the compound covariate methed under the Cox proportional hazard model.  
}
\description{
Calculate regression coefficients and their standard errors under the Cox proportional hazard model with
 the large number of covariates. 
}
\details{
\tabular{ll}{
Package: \tab compound.Cox\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-04-15\cr
License: \tab GPL-2\cr
}
}

\author{Takeshi Emura & Yi-Hau Chen}
\references{
T. Emura, Y.-H. Chen, H.-Y. Chen, Survival Prediction Based on Compound Covariate under Cox Proportional Hazard Models with Microarrays,
submitted to PloS ONE
}
\keyword{ package }
