\name{compound.Cox-package}
\alias{compound.Cox-package}
\alias{compound.Cox}
\docType{package}
\title{
Regression estimation based on the compound covariate method under the Cox proportional hazard model.  
}
\description{
Calculate regression coefficients and their standard deviations under the Cox proportional hazard model with the large number of covariates.
A shrinkage type estimation based on the compound covariate method (Emura, Chen & Chen, 2012 PLoS ONE) and a univariate gene selection method under dependent 
censoring (Emura & Chen, 2014 Statistical Methods in Medical Research) are implemented in this package. 
}
\details{
\tabular{ll}{
Package: \tab compound.Cox\cr
Type: \tab Package\cr
Version: \tab 1.4\cr
Date: \tab 2014-4-12\cr
License: \tab GPL-2\cr
}
}

\author{
Takeshi Emura & Yi-Hau Chen
Maintainer: Takeshi Emura <emura@stat.ncu.edu.tw>
}
\references{
Emura T, Chen Y-H, Chen H-Y (2012) Survival Prediction Based on Compound Covariate under Cox Proportional Hazard Models.
 PLoS ONE 7(10): e47627. doi:10.1371/journal.pone.0047627

Emura T and Chen YH (2014) Gene Selection for Survival Data Under Dependent Censoring: 
a Copula-based Approach, Statistical Methods in Medical Research, in press.
}
\keyword{ package }
