\name{compound.Cox-package}
\alias{compound.Cox-package}
\alias{compound.Cox}
\docType{package}
\title{
Regression estimation based on the compound shrinkage method under the Cox proportional hazard model.  
}
\description{
Calculate regression coefficients and their standard deviations under the Cox proportional hazard model with
 the large number of covariates. 
}
\details{
\tabular{ll}{
Package: \tab compound.Cox\cr
Type: \tab Package\cr
Version: \tab 1.3\cr
Date: \tab 2013-11-07\cr
License: \tab GPL-2\cr
}
}

\author{
Takeshi Emura & Yi-Hau Chen
Maintainer: Takeshi Emura <emura@stat.ncu.edu.tw>
}
\references{
Emura T, Chen Y-H, Chen H-Y (2012) Survival Prediction Based on Compound Covariate under Cox Proportional Hazard Models.
 PLoS ONE 7(10): e47627. doi:10.1371/journal.pone.0047627
}
\keyword{ package }
