\name{compound.Cox-package}
\alias{compound.Cox-package}
\alias{compound.Cox}
\docType{package}
\title{
Regression estimation based on the compound covariate method under the Cox proportional hazard model.  
}
\description{
Calculate regression coefficients and their standard deviations under the Cox proportional hazard model with
 the large number of covariates. 
}
\details{
\tabular{ll}{
Package: \tab compound.Cox\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2012-10-2\cr
License: \tab GPL-2\cr
}
}
\author{Takeshi Emura & Yi-Hau Chen}
\references{
Emura T, Chen YH, Chen HY (2012), Survival prediction based on compound covariate method under Cox proportional hazard models,
 PLoS ONE : e46362. doi:10.1371/journal.pone.0047627
}
\keyword{ package }