\name{compound.Cox-package}
\alias{compound.Cox-package}
\alias{compound.Cox}
\docType{package}
\title{
Regression Estimation Based on the Compound Covariate Method Under the Cox Proportional Hazard Model.  
}
\description{
Calculate regression coefficients and their standard deviations under the Cox proportional hazard model with a large number of covariates.
A shrinkage type estimation based on the compound covariate method (Emura, Chen & Chen, 2012 PLoS ONE) and a univariate gene selection method under dependent 
censoring (Emura & Chen, 2014 Statistical Methods in Medical Research) are implemented in this package. 
Algorithms for simulating gene expressions in these references are also given.
}
\details{
\tabular{ll}{
Package: \tab compound.Cox\cr
Type: \tab Package\cr
Version: \tab 2.1\cr
Date: \tab 2016-10-21\cr
License: \tab GPL-2\cr
}
}

\author{
Takeshi Emura & Yi-Hau Chen;
Maintainer: Takeshi Emura <takeshiemura@gmail.com>
}
\references{
Emura T, Chen Y-H, Chen H-Y (2012). Survival Prediction Based on Compound Covariate under Cox Proportional Hazard Models.
 PLoS ONE 7(10): e47627. doi:10.1371/journal.pone.0047627

Emura T, Chen YH (2014). Gene Selection for Survival Data Under Dependent Censoring: 
a Copula-based Approach, Statistical Methods in Medical Research, doi: 10.1177/0962280214533378.
}
\keyword{ package }
